Add basic drag and drop functionality
Dragged and dropped files now open in the app. To do this, I had
to make a wrapper over the QPlainTextEdit class to filter
drag and drop events.

The next step is to check the contents of the dragged files and
ask for user confirmation in cases that are unclear.